refactoring types
