+++ packages/ui/src/views/user-management/components/dialogs/index.ts
@@ -1,1 +1,2 @@
 export * from './dialogs'
+export * from './types'
