+++ packages/ui/src/views/user-management/components/page-components/content/content.tsx
@@ -1,12 +1,13 @@
 import { FiltersBar, PaginationComponent, Spacer } from '@/components'
 import { generateAlphaNumericHash } from '@/utils/utils'
 import { SandboxLayout } from '@/views'
+import { DialogLabels } from '@/views/user-management/components/dialogs'
 import { EmptyState } from '@/views/user-management/components/empty-state'
 import { Actions } from '@/views/user-management/components/page-components/actions'
 import { UsersList } from '@/views/user-management/components/page-components/content/components/users-list'
 import { ContentProps } from '@/views/user-management/components/page-components/content/types'
 import { Header } from '@/views/user-management/components/page-components/header'
-import { DialogLabels, UsersProps } from '@/views/user-management/types'
+import { UsersProps } from '@/views/user-management/types'
 import { getFilterOptions, getSortDirections, getSortOptions } from '@views/repo/constants/filter-options'
 import { useFilters } from '@views/repo/hooks'
 import { useDialogs } from '@views/user-management/providers/DialogsProvider'
