+++ packages/ui/src/views/user-management/components/tabs/types.ts
@@ -1,0 +1,4 @@
+export enum EActiveTab {
+  ACTIVE = 'active',
+  INACTIVE = 'inactive'
+}
