+++ packages/ui/src/views/user-management/providers/StoreProvider/index.ts
@@ -1,2 +1,3 @@
 export * from './hooks/useUserManagementStore'
 export * from './StoreProvider'
+export * from './types'
